Fix logic in conversion of mono audio into stereo
Recorded audio is at 2x speed when OWT sends mono audio.
This is because sizeInBytes is calculated based on the incoming
number_of_channels, and hence, even though the buffer holds
stereo data, only 1 channel is sent to vhal.

To fix this, sizeInBytes has to be calculated based on the
outgoing audio configuration: mAudioConfig.channelCount.

Tracked-On: OAM-105547
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>