ics3a/encoder: group irr_shared.so api in a dedicated header
This commit groups irr_shared.so transcoding API into 2 header
files: one for external usage (which is how we use it to build
icr_encoder service .exe) and for internal usage. Internal usage
API can be considered as a candidate to be moved to external
as soon someone will actually start using it.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>