ics3a/owt: upstream docker to build OWT
This PR provides Dockerfile to build OWT from sources. It corresponds
to the OWT prebuilt version we store in ./prebuilt/owt folder. Expectation
is that there is bit-to-bit match for the OWT distributable artifacts
wish someone to rebuild.

We still use prebuilt OWT version to build Streaming Service Container.
Reason is that this saves a lot for container build time. OWT checkout
process is both disk space and time consuming (fetches ~9GB of data and
typically takes ~10 or more minutes). Thus, it's just more convenient to
keep OWT build separate.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>