build: separate linux and windows parts in meson scripts
This commit separates linux and windows parts in meson scripts.
It suites as a preliminary step to allow building with meson
on windows. As of this commit windows build with meson is not
fully enabled, some dependencies are missing. However, work
with m4 docker templates should be possible both on Windows and
Linux after this commit:
  meson setup -Dm4=enabled -Dstreamer=disabled -Dencoder=disabled _build
  meson compile -C _build
  meson install -C _build

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>