win: upstream IDD driver and basic tools
This commit adds implementation for Indirect Display Driver (IDD)
customized to work with Intel GPUs in headless environment (i.e. on
ATS-M systems). Commit also adds idd-setup-tool for IDD driver
installation and system display settings customizations and
enum-adapters to enumerate available video adapters on the system.

That's the first commit in a series to upstream Windows streaming
support for headless setups with Intel GPUs. IDD driver should be
built directly via MSVC solution. The rest of the targets can be
built with meson. Currently it's required to explicitly disable
streamer component which is not yet upstreamed for Windows:

  meson setup -Dstreamer=disabled _build

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>