ics3a: simplify dockers after migration to meson
With unified build system covering all aspects of the project
we can simplify our dockerfiles since we now are capable to
reuse dependencies in a more flexible way.

This commit also catches that we had a header in owt directory
which actually is from older version of owt which we used it.
This header got deleted.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>