meson: rework dependencies declarations
This commit moves dependency declarations under the components
where these dependencies are used. This removes opprotunity to
see list of all dependencies of the project in one place. But
from the other hand this simplifies options handling in the project
and speeds up project build if some components are explicitly
disabled. The later is important for us because we have few
non-trivial dependencies such as owt and (to lesser extent) ffmpeg.
The list of all dependencies always can be just grepped:

  fgrep -rsn dependency | grep meson.build

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>