win: upstream webrtc-client
This commit adds Windows webrtc (owt) client application. It requires
a set of dependencies to be built first (in earlier commit we've
provided docker containers showing how to build them):

* ffmpeg w/ enabled DX11 decoders
* openssl
* owt (client build configuration)

Once dependencies are built, configure client as follows:
  meson setup _build --vsenv \
    --wrap-mode=forcefallback
    -Dprebuilt-path=/path/to/prebuilt/components \
    -Dstreamer=disabled

Note that --wrap-mode=forcefallback is required to fetch dependencies
defined on meson level. -Dstreamer=disabled is a current limitation
for the project build on Windows (streaming component is not yet
upstreamed).

This client can be used as a sample Windows client for Android CG.
It supports H.264, H265 and AV1 streaming.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>